@@ -1,181 +1,154 @@
 package com.salesflow.activity.service;
 
+import com.salesflow.activity.client.ContactCoreClient;
 import com.salesflow.activity.dto.ActivityDTO;
 import com.salesflow.activity.dto.TimelineDTO;
 import com.salesflow.activity.model.Activity;
 import com.salesflow.activity.model.ActivityType;
 import com.salesflow.activity.repository.ActivityRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.persistence.EntityNotFoundException;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class ActivityService {
     private final ActivityRepository activityRepository;
+    private final ContactCoreClient contactCoreClient;
 
     @Transactional
     public ActivityDTO createActivity(ActivityDTO activityDTO) {
-        Activity activity = convertToEntity(activityDTO);
+        // Verify if contact exists
+        ResponseEntity<Boolean> response = contactCoreClient.checkContactExists(activityDTO.getContactId());
+        if (!Boolean.TRUE.equals(response.getBody())) {
+            throw new EntityNotFoundException("Contact not found with ID: " + activityDTO.getContactId());
+        }
+
+        Activity activity = new Activity();
+        activity.setContactId(activityDTO.getContactId());
+        activity.setType(activityDTO.getType());
+        activity.setTitle(activityDTO.getTitle());
+        activity.setDescription(activityDTO.getDescription());
+        activity.setStatus(activityDTO.getStatus());
+        activity.setPriority(activityDTO.getPriority());
+        activity.setScheduledTime(activityDTO.getScheduledTime());
+        activity.setAssignedTo(activityDTO.getAssignedTo());
+        activity.setCreatedBy(activityDTO.getCreatedBy());
+        activity.setUpdatedBy(activityDTO.getUpdatedBy());
+
         Activity savedActivity = activityRepository.save(activity);
-        return convertToDTO(savedActivity);
+        return mapToDTO(savedActivity);
     }
 
     @Transactional(readOnly = true)
-    public ActivityDTO getActivity(Long id) {
+    public ActivityDTO getActivity(UUID id) {
         Activity activity = activityRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Activity not found with id: " + id));
-        return convertToDTO(activity);
+                .orElseThrow(() -> new EntityNotFoundException("Activity not found with ID: " + id));
+        return mapToDTO(activity);
     }
 
     @Transactional(readOnly = true)
-    public List<ActivityDTO> getActivitiesByContact(Long contactId) {
-        return activityRepository.findByContactId(contactId)
-                .stream()
-                .map(this::convertToDTO)
+    public List<ActivityDTO> getActivitiesByContact(UUID contactId) {
+        return activityRepository.findByContactId(contactId).stream()
+                .map(this::mapToDTO)
                 .collect(Collectors.toList());
     }
 
     @Transactional(readOnly = true)
-    public List<ActivityDTO> getActivitiesByType(Long contactId, ActivityType type) {
+    public List<ActivityDTO> getActivitiesByType(UUID contactId, ActivityType type) {
         return activityRepository.findByContactIdAndType(contactId, type)
                 .stream()
-                .map(this::convertToDTO)
+                .map(this::mapToDTO)
                 .collect(Collectors.toList());
     }
 
     @Transactional(readOnly = true)
     public List<ActivityDTO> getPendingActivities(String assignedTo) {
-        return activityRepository.findByAssignedToAndCompletedAtIsNull(assignedTo)
-                .stream()
-                .map(this::convertToDTO)
+        return activityRepository.findByAssignedToAndCompletedTimeIsNull(assignedTo).stream()
+                .map(this::mapToDTO)
                 .collect(Collectors.toList());
     }
 
     @Transactional
-    public ActivityDTO completeActivity(Long id) {
+    public ActivityDTO updateActivity(UUID id, ActivityDTO activityDTO) {
         Activity activity = activityRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Activity not found with id: " + id));
-        activity.setCompletedAt(LocalDateTime.now());
+                .orElseThrow(() -> new EntityNotFoundException("Activity not found with ID: " + id));
+
+        activity.setType(activityDTO.getType());
+        activity.setTitle(activityDTO.getTitle());
+        activity.setDescription(activityDTO.getDescription());
+        activity.setStatus(activityDTO.getStatus());
+        activity.setPriority(activityDTO.getPriority());
+        activity.setScheduledTime(activityDTO.getScheduledTime());
+        activity.setAssignedTo(activityDTO.getAssignedTo());
+        activity.setUpdatedBy(activityDTO.getUpdatedBy());
+
         Activity updatedActivity = activityRepository.save(activity);
-        return convertToDTO(updatedActivity);
+        return mapToDTO(updatedActivity);
     }
 
     @Transactional
-    public ActivityDTO updateActivity(Long id, ActivityDTO activityDTO) {
-        Activity existingActivity = activityRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Activity not found with id: " + id));
-        
-        updateEntityFromDTO(existingActivity, activityDTO);
-        Activity updatedActivity = activityRepository.save(existingActivity);
-        return convertToDTO(updatedActivity);
+    public void deleteActivity(UUID id) {
+        if (!activityRepository.existsById(id)) {
+            throw new EntityNotFoundException("Activity not found with ID: " + id);
+        }
+        activityRepository.deleteById(id);
     }
 
     @Transactional(readOnly = true)
-    public List<TimelineDTO> getContactTimeline(Long contactId, LocalDateTime startDate, LocalDateTime endDate) {
-        List<Activity> activities = activityRepository.findByContactId(contactId);
-        
-        return activities.stream()
-                .filter(activity -> {
-                    if (startDate != null && activity.getScheduledAt().isBefore(startDate)) {
-                        return false;
-                    }
-                    if (endDate != null && activity.getScheduledAt().isAfter(endDate)) {
-                        return false;
-                    }
-                    return true;
-                })
-                .map(this::convertToTimelineDTO)
-                .collect(Collectors.toList());
+    public TimelineDTO getContactTimeline(UUID contactId, LocalDateTime startDate) {
+        List<Activity> activities = activityRepository.findByContactIdAndDateRange(contactId, startDate, LocalDateTime.now());
+        TimelineDTO timelineDTO = new TimelineDTO();
+        timelineDTO.setContactId(contactId);
+        timelineDTO.setActivities(activities.stream()
+                .map(this::mapToDTO)
+                .collect(Collectors.toList()));
+        return timelineDTO;
     }
 
     @Transactional(readOnly = true)
-    public List<ActivityDTO> searchActivities(String query, ActivityType type, String assignedTo, 
+    public List<ActivityDTO> searchActivities(String query, ActivityType type, String assignedTo,
                                             LocalDateTime startDate, LocalDateTime endDate) {
-        List<Activity> activities = activityRepository.findAll();
-        
-        return activities.stream()
-                .filter(activity -> {
-                    if (query != null && !query.isEmpty()) {
-                        if (!activity.getNotes().toLowerCase().contains(query.toLowerCase()) &&
-                            !activity.getOutcome().toLowerCase().contains(query.toLowerCase())) {
-                            return false;
-                        }
-                    }
-                    if (type != null && activity.getType() != type) {
-                        return false;
-                    }
-                    if (assignedTo != null && !assignedTo.equals(activity.getAssignedTo())) {
-                        return false;
-                    }
-                    if (startDate != null && activity.getScheduledAt().isBefore(startDate)) {
-                        return false;
-                    }
-                    if (endDate != null && activity.getScheduledAt().isAfter(endDate)) {
-                        return false;
-                    }
-                    return true;
-                })
-                .map(this::convertToDTO)
+        return activityRepository.searchActivities(query, type, assignedTo, startDate, endDate).stream()
+                .map(this::mapToDTO)
                 .collect(Collectors.toList());
     }
 
-    private Activity convertToEntity(ActivityDTO dto) {
-        Activity activity = new Activity();
-        activity.setContactId(dto.getContactId());
-        activity.setType(dto.getType());
-        activity.setOutcome(dto.getOutcome());
-        activity.setNotes(dto.getNotes());
-        activity.setScheduledAt(dto.getScheduledAt());
-        activity.setCompletedAt(dto.getCompletedAt());
-        activity.setAssignedTo(dto.getAssignedTo());
-        activity.setNextStep(dto.getNextStep());
-        return activity;
+    @Transactional
+    public ActivityDTO completeActivity(UUID id) {
+        Activity activity = activityRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Activity not found with ID: " + id));
+
+        activity.setStatus("COMPLETED");
+        activity.setCompletedTime(LocalDateTime.now());
+
+        Activity completedActivity = activityRepository.save(activity);
+        return mapToDTO(completedActivity);
     }
 
-    private ActivityDTO convertToDTO(Activity activity) {
+    private ActivityDTO mapToDTO(Activity activity) {
         ActivityDTO dto = new ActivityDTO();
         dto.setId(activity.getId());
         dto.setContactId(activity.getContactId());
         dto.setType(activity.getType());
-        dto.setOutcome(activity.getOutcome());
-        dto.setNotes(activity.getNotes());
-        dto.setScheduledAt(activity.getScheduledAt());
-        dto.setCompletedAt(activity.getCompletedAt());
+        dto.setTitle(activity.getTitle());
+        dto.setDescription(activity.getDescription());
+        dto.setStatus(activity.getStatus());
+        dto.setPriority(activity.getPriority());
+        dto.setScheduledTime(activity.getScheduledTime());
+        dto.setCompletedTime(activity.getCompletedTime());
         dto.setAssignedTo(activity.getAssignedTo());
-        dto.setNextStep(activity.getNextStep());
+        dto.setCreatedBy(activity.getCreatedBy());
+        dto.setUpdatedBy(activity.getUpdatedBy());
         dto.setCreatedAt(activity.getCreatedAt());
         dto.setUpdatedAt(activity.getUpdatedAt());
         return dto;
-    }
-
-    private TimelineDTO convertToTimelineDTO(Activity activity) {
-        TimelineDTO dto = new TimelineDTO();
-        dto.setId(activity.getId());
-        dto.setContactId(activity.getContactId());
-        dto.setType(activity.getType());
-        dto.setOutcome(activity.getOutcome());
-        dto.setNotes(activity.getNotes());
-        dto.setTimestamp(activity.getScheduledAt());
-        dto.setAssignedTo(activity.getAssignedTo());
-        dto.setNextStep(activity.getNextStep());
-        dto.setCompleted(activity.getCompletedAt() != null);
-        dto.setCompletedAt(activity.getCompletedAt());
-        return dto;
-    }
-
-    private void updateEntityFromDTO(Activity activity, ActivityDTO dto) {
-        if (dto.getType() != null) activity.setType(dto.getType());
-        if (dto.getOutcome() != null) activity.setOutcome(dto.getOutcome());
-        if (dto.getNotes() != null) activity.setNotes(dto.getNotes());
-        if (dto.getScheduledAt() != null) activity.setScheduledAt(dto.getScheduledAt());
-        if (dto.getCompletedAt() != null) activity.setCompletedAt(dto.getCompletedAt());
-        if (dto.getAssignedTo() != null) activity.setAssignedTo(dto.getAssignedTo());
-        if (dto.getNextStep() != null) activity.setNextStep(dto.getNextStep());
     }
 } 
