@@ -6,54 +6,38 @@ import com.salesflow.activity.model.ActivityType;
 import com.salesflow.activity.repository.ActivityRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Duration;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 public class ActivityStatsService {
     private final ActivityRepository activityRepository;
 
-    public ActivityStatsDTO getContactActivityStats(Long contactId) {
+    @Transactional(readOnly = true)
+    public ActivityStatsDTO getActivityStats(UUID contactId) {
         List<Activity> activities = activityRepository.findByContactId(contactId);
         
         int totalActivities = activities.size();
         int completedActivities = (int) activities.stream()
-                .filter(a -> a.getCompletedAt() != null)
+                .filter(a -> a.getCompletedTime() != null)
                 .count();
-        int pendingActivities = totalActivities - completedActivities;
         
-        Map<ActivityType, Integer> activitiesByType = activities.stream()
-                .collect(Collectors.groupingBy(
-                        Activity::getType,
-                        Collectors.collectingAndThen(Collectors.counting(), Long::intValue)
-                ));
+        ActivityStatsDTO stats = new ActivityStatsDTO();
+        stats.setContactId(contactId);
+        stats.setTotalActivities(totalActivities);
+        stats.setCompletedActivities(completedActivities);
         
-        double completionRate = totalActivities > 0 
-                ? (double) completedActivities / totalActivities * 100 
-                : 0;
-        
-        int averageResponseTime = calculateAverageResponseTime(activities);
-        
-        return new ActivityStatsDTO(
-                contactId,
-                totalActivities,
-                completedActivities,
-                pendingActivities,
-                activitiesByType,
-                completionRate,
-                averageResponseTime
-        );
+        return stats;
     }
 
     private int calculateAverageResponseTime(List<Activity> activities) {
         return (int) activities.stream()
-                .filter(a -> a.getCompletedAt() != null && a.getScheduledAt() != null)
-                .mapToLong(a -> Duration.between(a.getScheduledAt(), a.getCompletedAt()).toHours())
+                .filter(a -> a.getCompletedTime() != null && a.getScheduledTime() != null)
+                .mapToLong(a -> Duration.between(a.getScheduledTime(), a.getCompletedTime()).toHours())
                 .average()
                 .orElse(0);
     }
